refactor: post, user comment 기능 추가, user patch(edit user) 기능 추가

+++ back/src/main/java/com/demo/library/post/controller/PostController.java
@@ -4,6 +4,7 @@ import com.demo.library.book.dto.BookDto;
 import com.demo.library.book.entity.BookEntity;
 import com.demo.library.book.mapper.BookMapper;
 import com.demo.library.book.service.BookService;
+import com.demo.library.exception.BusinessLogicException;
 import com.demo.library.post.dto.PostDto;
 import com.demo.library.post.entity.PostEntity;
 import com.demo.library.post.mapper.PostMapper;
@@ -23,6 +24,9 @@ import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
+
+import static com.demo.library.exception.ExceptionCode.NOT_ALLOWED;
 
 @RestController
 @RequestMapping("/posts")
@@ -63,6 +67,16 @@ public class PostController {
 
         return ResponseCreator.list(listedPosts,postPage.getTotalPages());
     }
+    @GetMapping("/book-posts/{book-id}")
+    public ResponseEntity<ListResponseDto<PostDto.Listed>>
+                                get(Pageable pageable,@PathVariable("book-id") Long Id) {
+        Page<PostEntity> postPage = service.getPostsByBookId(pageable,Id);
+        List<PostEntity> postList = postPage.getContent();
+
+        List<PostDto.Listed> listedPosts = mapper.postsToList(postList);
+
+        return ResponseCreator.list(listedPosts,postPage.getTotalPages());
+    }
 
     @GetMapping("/{post-id}")
     public ResponseEntity<SingleResponseDto<PostDto.Response>>
@@ -70,12 +84,21 @@ public class PostController {
 
         PostEntity postEntity = service.getPost(Id);
         PostDto.Response responseDto = mapper.postEntityToResponse(postEntity);
+
+        Boolean isAuthor = Optional.ofNullable(jwtAuthService.getEmail())
+                .map(email -> service.checkIsAuthor(postEntity, userService.findByEmail(email)))
+                .orElse(false);
+
+        responseDto.setIsAuthor(isAuthor);
         return ResponseCreator.single(responseDto);
     }
     @DeleteMapping("/{post-id}")
     public ResponseEntity<Void> delete(@PathVariable("post-id") Long Id) {
 
-        service.deletePost(Id,userService.findByEmail(jwtAuthService.getEmail()));
+        PostEntity postEntity = service.getPost(Id);
+        if(service.checkIsAuthor(postEntity,userService.findByEmail(jwtAuthService.getEmail())))
+            service.deletePost(Id,userService.findByEmail(jwtAuthService.getEmail()));
+        else throw new BusinessLogicException(NOT_ALLOWED);
 
         return ResponseCreator.deleted();
     }
